# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
...
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
    Role:
      type: string
      enum:
        - ADMIN
        - USER1
        - USER2
        - INACTIVE
      example: USER
    Product:
      type: object
      required: 
        - barcode
        - name
        - category
        - sellPrice
        - stock
      properties: 
        barcode: 
          type: string
          example: "6430039222212"
          description: Barcode on the product packaging. Used to uniquely identify the product.
        name: 
          type: string
          example: Kombucha bakteerin makuinen bakteerijuoma
          description: Human readable and user facing name of the product.
        category: 
          '$ref': '#/components/schemas/Category'
        sellPrice: 
          type: integer
          example: 310
          description: Price (in cents) at which the product is on sale.  
        stock: 
          type: integer
          example: 10
          description: Number of items in stock.
    AdminProduct:
      allOf:
        - '$ref': '#/components/schemas/Product'
        - type: object
          required:
            - buyPrice
          properties:
            buyPrice:
              type: integer
              example: 300
              description: Price (in cents) at which the product was bought in.
...
    PurchaseEventBase:
      type: object
      required: 
        - purchaseId
        - time
        - price
      properties: 
        purchaseId: 
          type: integer
          example: 3053
          description: Opaque unique identifier of the purchase event.
        time: 
          type: string
          format: date-time
          example: '2018-12-24T00:00:00.000Z'
          description: Timestamp of the event in ISO-8601 format.
        price: 
          type: integer
          example: 19
          description: Price (in cents) of the product bought.
...
    PurchaseEventUser:
      type: object
      required:
        - user
      properties:
        user:
          '$ref': '#/components/schemas/User'
    PurchaseEventBalance:
      type: object
      required:
        - balanceAfter
      properties:
        balanceAfter:
          type: integer
          example: 1981
          description: Balance of the user's account (in cents) after the purchase transaction.
    PurchaseEventStock:
      type: object
      required:
        - stockAfter
      properties:
        stockAfter:
          type: integer
          example: 11
          description: Number of the bought product left in stock after the purchase transaction.
...
    User: 
      type: object
      required: 
        - userId
        - username
        - email
        - moneyBalance
        - role
        - privacyLevel
      properties: 
        userId: 
          type: integer
          example: 415
          description: Opaque unique identifier of the user.
        username: 
          type: string
          example: fuxi2001
          description: Username which is used in authentication.
        fullName:
          type: string
          nullable: true
          example: Fuxi C. S. McStudent
          description: Real full name of the user.
        email: 
          type: string
          example: fuxi.student@helsinki.fi
          description: E-mail address at which the user can be reached.
        moneyBalance: 
          type: integer
          example: 2000
          description: Balance of the user's account in cents.
        privacyLevel:
          type: integer
          example: 0
          description: Privacy level of the user, 0 = no restrictions, 1 = hide username, 2 = hide even anonymous data
          enum: [0, 1, 2]
        role:
          '$ref': '#/components/schemas/Role'
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
          description: Opaque unique identifier of the category.
        description: 
          type: string
          example: Drinks, soda with bottle return
          description: Human readable and user facing description of the product category.
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
                  - invalid_token
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/products: 
    get: 
      summary: Get a list of all products
      operationId: getProducts
      tags:
        - normal
      security:
        - normalAccessToken: []
      responses: 
        200:
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - products
                properties: 
                  products: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Product'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/products/{barcode}: 
    get:
      summary: Get details of a specific product
      operationId: getProduct
      tags:
        - normal
      security:
        - normalAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema:
                type: object
                required: 
                  - product
                properties: 
                  product:
                    '$ref': '#/components/schemas/Product'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/products/{barcode}/purchase: 
    post:
      summary: Purchase a prouct
      operationId: purchase
      tags:
        - normal
      security:
        - normalAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200:
          description: Purchase successfull
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - accountBalance
                  - productStock
                  - purchases
                properties: 
                  accountBalance: 
                    type: integer
                    example: 19.81
                  productStock: 
                    type: integer
                    example: 11
                  purchases: 
                    type: array
                    items:
                      allOf:
                        - '$ref': '#/components/schemas/PurchaseEventBase'
                        - '$ref': '#/components/schemas/PurchaseEventBalance'
                        - '$ref': '#/components/schemas/PurchaseEventStock'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        403:
          description: The account has insufficient funds and the purchase cannot be made.
          content:
            application/json:
              schema:
                type: object
                required:
                  - error_code
                  - message
                properties:
                  error_code:
                    type: string
                    enum:
                      - insufficient_funds
                  message:
                    type: string
                    example: Insufficient funds
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - count
              properties: 
                count: 
                  type: integer
                  example: 1
                  minimum: 1
...
  /api/v1/admin/products: 
    get: 
      summary: Get a list of all products
      operationId: getProducts
      tags:
        - admin
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - products
                properties: 
                  products: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    post: 
      summary: Create a new product
      operationId: createProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      responses: 
        201: 
          description: New product created successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - barcode
                - name
                - categoryId
                - buyPrice
                - sellPrice
                - stock
              properties: 
                barcode: 
                  type: string
                  example: '7314660003114'
                  minLength: 1
                  maxLength: 14
                name: 
                  type: string
                  example: Muumilimsa
                  minLength: 1
                categoryId: 
                  type: integer
                  example: 10
                  minimum: 0
                buyPrice: 
                  type: integer
                  example: 210
                sellPrice: 
                  type: integer
                  example: 220
                stock: 
                  type: integer
                  example: 8
  /api/v1/admin/products/{barcode}: 
    get: 
      summary: Get details of a specific product
      operationId: getProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update the details of a specific product
      operationId: updateProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              minProperties: 1
              additionalProperties: false
              properties: 
                name: 
                  type: string
                  example: Muumilimsa
                  minLength: 1
                categoryId: 
                  type: integer
                  example: 10
                  minimum: 0
                buyPrice: 
                  type: integer
                  example: 210
                sellPrice: 
                  type: integer
                  example: 220
                stock: 
                  type: integer
                  example: 8
    delete: 
      summary: Delete a product
      operationId: deleteProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedProduct
                properties: 
                  deletedProduct: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/products/{barcode}/purchaseHistory: 
    get: 
      summary: Get a list of product's purchase events
      operationId: getProductPurchaseHistory
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - purchases
                properties: 
                  purchases: 
                    type: array
                    items:
                      allOf:
                        - '$ref': '#/components/schemas/PurchaseEventBase'
                        - '$ref': '#/components/schemas/PurchaseEventUser'
                        - '$ref': '#/components/schemas/PurchaseEventStock'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...